pyproject.toml
[tool.poetry]
name = "python-api-full"
version = "0.1.0"
description = ""
authors = ["Davide"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = {extras = ["all"], version = "^0.70.0"}
pydantic = "^1.8.2"
psycopg2-binary = "^2.9.2"
SQLAlchemy = "^1.4.27"
email-validator = "^1.1.3"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
alembic = "^1.7.5"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
black = {version = "^21.10b0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
